requirements.txt
numpy>=1.19.1
requests>=2.20.0
matplotlib>=3.5.0
python-dateutil>=2.8
pyocr
Pillow
openpyxl
pandas
python-docx
wxPython
plyvel
pytz
